Move to extensions the functions that will not be overwritten but that may be used on the elements
