+++ godtools/Views/TractElements/Number.swift
@@ -12,9 +12,10 @@ import UIKit
 class Number: BaseTractElement {
     
     let paddingConstant = CGFloat(30.0)
+    static let widthConstant = CGFloat(50.0)
     
     override func setupView(properties: Dictionary<String, Any>) {
-        let frame = CGRect(x: 0.0, y: self.yStartPosition  + paddingConstant, width: BaseTractElement.Standards.screenWidth, height: self.height)
+        let frame = CGRect(x: 0.0, y: self.yStartPosition  + paddingConstant, width: Number.widthConstant, height: self.height)
         let view = UIView(frame: frame)
         view.backgroundColor = .green
         self.view = view
